clients/upsclient.h and numerous *.c: align upscli_get(), upscli_list_start(), upscli_list_next() numa and numq args from "unsigned int" to "size_t", to match PCONF_CTX_t fields

+++ clients/upscmd.c
@@ -65,7 +65,7 @@ static void usage(const char *prog)
 static void print_cmd(char *cmdname)
 {
 	int		ret;
-	unsigned int	numq, numa;
+	size_t	numq, numa;
 	const char	*query[4];
 	char		**answer;
 
@@ -88,7 +88,7 @@ static void print_cmd(char *cmdname)
 static void listcmds(void)
 {
 	int		ret;
-	unsigned int	numq, numa;
+	size_t	numq, numa;
 	const char	*query[4];
 	char		**answer;
 	struct list_t	*lhead = NULL, *llast = NULL, *ltmp, *lnext;
@@ -113,7 +113,7 @@ static void listcmds(void)
 
 		/* CMD <upsname> <cmdname> */
 		if (numa < 3) {
-			fatalx(EXIT_FAILURE, "Error: insufficient data (got %d args, need at least 3)", numa);
+			fatalx(EXIT_FAILURE, "Error: insufficient data (got %zu args, need at least 3)", numa);
 		}
 
 		/* we must first read the entire list of commands,
